- компоненты вынесены по отдельным файлам - переменные вынесены под импорты
