+++ pkg/builder/builder_test.go
@@ -5,10 +5,10 @@ import (
 	"testing"
 )
 
-func TestBuilder(t *testing.T) {
-	toys, bed, floor := "all", "accurately", "carefully"
-	expect := fmt.Sprintf("Collected %v toys\nMade the bed %v\nVacuumed the floor %v\n", toys, bed, floor)
+var toys, bed, floor = "all", "accurately", "carefully"
 
+func TestBuilder(t *testing.T) {
+	expect := fmt.Sprintf("Collected %v toys\nMade the bed %v\nVacuumed the floor %v\n", toys, bed, floor)
 	order := NewOrder()
 	mother := NewMother(NewCleanUpTheRoom(order))
 	mother.Cleaning(toys, bed, floor)
